Redo project structure
