+++ src/Component.py
@@ -1,0 +1,5 @@
+from Node import Node
+
+class Component(Node):
+    def __init__(self, id):
+        super(Component, self).__init__(id)
