+++ src/Junction.py
@@ -1,0 +1,9 @@
+from Node import Node
+
+class Junction(Node):
+    def __init__(self, id):
+        super(Junction, self).__init__(id)
+        self.connections = [] # array of nodes
+
+    def add_connection(self, node):
+        self.connections.append(node)
